+++ tests/Unit/DependencyInjection/ConfigurationTest.php
@@ -47,6 +47,7 @@ final class ConfigurationTest extends TestCase
         self::assertSame('', $config['comments']['captcha']['site_key']);
         self::assertSame(0.5, $config['comments']['captcha']['min_score']);
         self::assertSame('website', $config['comments']['captcha']['honeypot_field']);
+        self::assertSame(5.0, $config['comments']['captcha']['timeout_seconds']);
         self::assertSame('none', $config['html']['sanitize']['strategy']);
         self::assertNull($config['html']['sanitize']['service']);
     }